organize folder for markdown decorators
