+++ src/components/blog-post-card/index.ts
@@ -1,3 +1,2 @@
 export * from './BlogPostCard';
 export * from './PropTypes';
-export * from './Decorator';
