+++ src/components/button-link/index.ts
@@ -1,0 +1,2 @@
+export * from './ButtonLink';
+export * from './PropTypes';
