+++ src/components/project-card/index.ts
@@ -1,3 +1,2 @@
 export * from './ProjectCard';
-export * from './Decorator';
 export * from './PropTypes';
